assignment/requirement.txt
asgiref==3.2.10
cffi==1.14.3
cryptography==3.1.1
Django==3.1.2
django-rest-knox==4.1.0
django-sendgrid-v5==0.9.0
djangorestframework==3.12.1
future==0.18.2
login==0.0.6
pycparser==2.20
python-http-client==3.3.1
pytz==2020.1
sendgrid==6.4.7
six==1.15.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
